refactored company service
