+++ frontend/frontendAngular/src/app/app.module.ts
@@ -18,7 +18,6 @@ import { MatNativeDateModule } from '@angular/material/core';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { MessageComponent } from './components/message/message.component';
 import { LoginComponent } from './components/login/login.component';
 import { RegisterCompanyComponent } from './components/register-company/register-company.component';
 import { RegisterCustomerComponent } from './components/register-customer/register-customer.component';
@@ -35,7 +34,6 @@ import { DialogComponent } from './components/dialog/dialog.component';
 @NgModule({
   declarations: [
     AppComponent,
-    MessageComponent,
     LoginComponent,
     RegisterCompanyComponent,
     RegisterCustomerComponent,
