--- frontend/frontendAngular/src/app/components/message/message.component.scss
@@ -1,20 +1,0 @@
-/* MessagesComponent's private CSS styles */
-h2 {
-  color: #A80000;
-  font-family: Arial, Helvetica, sans-serif;
-  font-weight: lighter;
-}
-
-.clear {
-  color: #333;
-  background-color: #eee;
-  margin-bottom: 12px;
-  padding: 1rem;
-  border-radius: 4px;
-  font-size: 1rem;
-}
-
-.clear:hover {
-  color: white;
-  background-color: #42545C;
-}
