--- frontend/frontendAngular/src/app/components/message/message.component.ts
@@ -1,16 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { MessageService } from 'src/app/services/message.service';
-
-@Component({
-  selector: 'app-message',
-  templateUrl: './message.component.html',
-  styleUrls: ['./message.component.scss']
-})
-export class MessageComponent implements OnInit {
-
-  constructor(public messageService: MessageService) { }
-
-  ngOnInit(): void {
-  }
-
-}
