+++ frontend/frontendAngular/src/app/components/register-company/register-company.component.scss
@@ -1,3 +1,3 @@
 mat-form-field {
-  height: 4.6em;
+  height: 4.45em;
 }
