+++ frontend/frontendAngular/src/app/components/user-home/user-home.component.html
@@ -1,36 +1,33 @@
 <div class="col-12">
 
-  <div class="row justify-content-between my-5 mx-5">
+  <div class="row justify-content-around">
+        <div class="row justify-content-start">
+          <div class="col-10">
+            <input #searchBox [(ngModel)]="result" (input)="search(searchBox.value)" placeholder="Pesquisar" [matAutocomplete]="auto">
+            <mat-autocomplete #auto="matAutocomplete">
+              <mat-option *ngFor="let card of cards$ | async" [value]="card.name">
+                {{card.name}}
+              </mat-option>
+            </mat-autocomplete>
+          </div>
+          <button id="search" mat-icon-button class="search-button" (click)="updateCards()">
+            <mat-icon>search</mat-icon>
+          </button>
+        </div>
 
-    <div class="col-6 ml-3">
-      
-      <div class="col-10">
-        <input #searchBox [(ngModel)]="result" (input)="search(searchBox.value)" placeholder="Pesquisar" [matAutocomplete]="auto">
-        <mat-autocomplete #auto="matAutocomplete">
-          <mat-option *ngFor="let card of cards$ | async" [value]="card.name">
-            {{card.name}}
-          </mat-option>
-        </mat-autocomplete>
+
+      <div class="col-3">
+        <mat-form-field appearance="fill">
+          <mat-label>Categorias</mat-label>
+          <mat-select [(ngModel)]="category">
+            <mat-option value="" (click)="updateCategory()">Geral</mat-option>
+            <mat-option *ngFor="let category of categories" [value]="category.value" (click)="updateCategory()">
+              {{category.viewValue}}
+            </mat-option>
+          </mat-select>
+        </mat-form-field>
       </div>
 
-      <div class="col-2">
-        <button id="search" mat-icon-button class="search-button" (click)="updateCards()">
-          <mat-icon>search</mat-icon>
-        </button>
-      </div>
-    </div>
-
-    <div class="col-4  mr-5">
-      <mat-form-field appearance="fill">
-        <mat-label>Categorias</mat-label>
-        <mat-select [(ngModel)]="category">
-          <mat-option value="" (click)="updateCategory()">Geral</mat-option>
-          <mat-option *ngFor="let category of categories" [value]="category.value" (click)="updateCategory()">
-            {{category.viewValue}}
-          </mat-option>
-        </mat-select>
-      </mat-form-field>
-    </div>
   </div>
 
   <div class="col-11">
@@ -53,3 +50,58 @@
     </div>
   </div>
 </div>
+
+<!-- <div class="col-12">
+  <div class="row justify-content-between my-5 mx-5">
+    <div class="row justify-content-start ml-3">
+      <div class="col-10">
+        <input #searchBox [(ngModel)]="result" (input)="search(searchBox.value)" placeholder="Pesquisar">
+        <ul class="search-result">
+          <li *ngFor="let card of cards$ | async">
+            <a routerLink="/detail/{{card.id}}">
+              {{card.name}}
+            </a>
+          </li>
+        </ul>
+      </div>
+
+      <div class="col-2">
+        <button id="search" mat-icon-button class="search-button" (click)="updateCards()">
+          <mat-icon>search</mat-icon>
+        </button>
+      </div>
+    </div>
+
+    <div class="justify-content-end mr-5">
+      <mat-form-field appearance="fill">
+        <mat-label>Categorias</mat-label>
+        <mat-select [(ngModel)]="category">
+          <mat-option value="" (click)="updateCategory()">Geral</mat-option>
+          <mat-option *ngFor="let category of categories" [value]="category.value" (click)="updateCategory()">
+            {{category.viewValue}}
+          </mat-option>
+        </mat-select>
+      </mat-form-field>
+    </div>
+  </div>
+
+  <div class="col-11">
+    <div class="row justify-content-around">
+      <div class="col-6 my-3" *ngFor="let card of cards">
+        <button class="col-12 ml-5 card-button" (click)="OpenDialog(card)">
+          <img src="assets/SVGCompany_1.svg">
+          <h2>{{card.name}}</h2>
+          <h3>Serviço: {{card.category}}</h3>
+          <p>
+            Cidade: {{card.city}} <br>
+            UF: {{card.uf}} <br>
+            Endereço: {{card.address}} <br>
+            CEP: {{card.cep}} <br>
+            CNPJ: {{card.cnpj}} <br>
+            Telefone: {{card.telephone_number}}
+          </p>
+        </button>
+      </div>
+    </div>
+  </div>
+</div> -->
